Applied even more warning flags!
Fixing all the warnings wasn't that easy actually..

+++ kernel/features/syscall.c
@@ -41,7 +41,7 @@ static void syscall_handler(struct regs *r)
 		if (vfs_ready((char *)r->ebx)) {
 			r->eax = (u32)vfs_read((char *)r->ebx, (void *)r->ecx, r->edx, r->esi);
 		} else {
-			if (vfs_wait((char *)r->ebx, vfs_read) < 0)
+			if (vfs_wait((char *)r->ebx, (u32)vfs_read) < 0)
 				r->eax = -1;
 			else
 				proc_yield(r);
